[hotfix #129] albumCover타입 Long으로 변경

+++ src/main/java/com/pophory/pophoryserver/domain/member/dto/request/MemberCreateV2RequestDto.java
@@ -28,7 +28,7 @@ public class MemberCreateV2RequestDto {
     private String nickname;
 
     @Schema(description = "앨범 커버 1 ~ 4 중의 값으로 요청", example = "1")
-    private int albumCover;
+    private Long albumCover;
 
     @Schema(description = "멤버의 fcmToken")
     private String fcmToken;
